backend/requirements.txt
fastapi
uvicorn
autogen
gitpython
pytest
pygithub
python-dotenv
openai
requests